Correcciones

+++ backend/controllers/product.controller.js
@@ -44,11 +44,14 @@ export const updateProduct = async(req, res) => {
 
 export const deleteProduct = async(req, res) => {
   const { id } = req.params;
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return res.status(404).json({message: " Product Not Found"});
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({message: "Product deleted"});
   } catch (error) {
     console.error("Error: ", error.message);
-    res.status(404).json({message: "Product not found"});
+    res.status(500).json({message: "Server Error"});
   }
 }
